Create README.md
The object of This beauty parlour System will be a web based 
system which will resolve the traditional problem of maintaining 
client’ records and appointment making procedure. As the 
numbers of web surfers are increasing day by day, a web base 
solution will be certainly more accurate than a traditional desktop 
base solution. The system will have different level of user. The 
one who is not registered as a user will only be able to view the 
basic information page. The users will be1. System Administration 
2. Registered user 
3. user 
The list of operations that the system will provide are1. Providing a client ID to the registered 
users. 
2. Maintaining records such as registered client. 
3. Registred client can get suggestion from the 
beauty expert. 
4. Users can give the review about their work. 
5. Users can see the work of the beauty parlor. 
6. Allowing the registered user to get an appointment 
7. customer can contract to the manager for any 
query